week02/day13/day13_faiss_gradio_hnsw_csv_space/requirements.txt
faiss-cpu
numpy
pandas
matplotlib
gradio
